+++ src/entities/Movimentacao.java
@@ -1,5 +1,4 @@
 package entities;
-
 import java.util.Date;
 
 public class Movimentacao {
